Refactor for wso2-extension to siddhi-io repo move and sync with siddhi 5.0.0
