+++ component/pom.xml
@@ -18,9 +18,9 @@
   -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <groupId>org.wso2.extension.siddhi.store.elasticsearch</groupId>
+        <groupId>io.siddhi.extension.store.elasticsearch</groupId>
         <artifactId>siddhi-store-elasticsearch-parent</artifactId>
-        <version>2.0.1-SNAPSHOT</version>
+        <version>3.0.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>siddhi-store-elasticsearch</artifactId>
@@ -28,19 +28,19 @@
     <name>Siddhi Extension - Elasticsearch Store Implementation</name>
     <dependencies>
         <dependency>
-            <groupId>org.wso2.siddhi</groupId>
+            <groupId>io.siddhi</groupId>
             <artifactId>siddhi-core</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.siddhi</groupId>
+            <groupId>io.siddhi</groupId>
             <artifactId>siddhi-annotations</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.siddhi</groupId>
+            <groupId>io.siddhi</groupId>
             <artifactId>siddhi-query-compiler</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.wso2.siddhi</groupId>
+            <groupId>io.siddhi</groupId>
             <artifactId>siddhi-query-api</artifactId>
         </dependency>
         <dependency>
@@ -98,7 +98,7 @@
             <build>
                 <plugins>
                     <plugin>
-                        <groupId>org.wso2.siddhi</groupId>
+                        <groupId>io.siddhi</groupId>
                         <artifactId>siddhi-doc-gen</artifactId>
                         <version>${siddhi.version}</version>
                         <executions>
@@ -195,6 +195,19 @@
     <build>
         <plugins>
             <plugin>
+                <groupId>io.siddhi</groupId>
+                <artifactId>siddhi-doc-gen</artifactId>
+                <version>${siddhi.version}</version>
+                <executions>
+                    <execution>
+                        <phase>compile</phase>
+                        <goals>
+                            <goal>generate-md-docs</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
                 <groupId>org.jacoco</groupId>
                 <artifactId>jacoco-maven-plugin</artifactId>
                 <executions>
@@ -251,8 +264,8 @@
                             org.apache.commons.codec.*,
                         </Private-Package>
                         <Export-Package>
-                            org.wso2.extension.siddhi.store.elasticsearch,
-                            org.wso2.extension.siddhi.store.elasticsearch.*
+                            io.siddhi.extension.store.elasticsearch,
+                            io.siddhi.extension.store.elasticsearch.*
                         </Export-Package>
                         <Import-Package>
                             org.elasticsearch.*;version="${elasticsearch.version.range}",
@@ -261,9 +274,9 @@
                             org.apache.logging.log4j.*,
                             org.apache.commons.logging.*,
                             org.joda.time.*,
-                            org.wso2.siddhi.core.*;version="${siddhi.version.range}",
-                            org.wso2.siddhi.annotation.*;version="${siddhi.version.range}",
-                            org.wso2.siddhi.query.api.*;version="${siddhi.version.range}",
+                            io.siddhi.core.*;version="${siddhi.version.range}",
+                            io.siddhi.annotation.*;version="${siddhi.version.range}",
+                            io.siddhi.query.api.*;version="${siddhi.version.range}",
                             org.apache.log4j.*,
                             com.fasterxml.jackson.core.*,
                             com.carrotsearch.hppc.*,
