+++ docs/index.md
@@ -42,7 +42,7 @@ Latest API Docs is <a target="_blank" href="https://wso2-extensions.github.io/si
 
 ```
      <dependency>
-        <groupId>org.wso2.extension.siddhi.store.elasticsearch</groupId>
+        <groupId>io.siddhi.extension.store.elasticsearch</groupId>
         <artifactId>siddhi-store-elasticsearch</artifactId>
         <version><version>x.x.x</version></version>
      </dependency>
